pyproject.toml
[project]
name = "fontana"
version = "0.1.0"
description = ""
authors = [{ name = "Nhu Viet Nguyen", email = "braveryandglory@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13"

# Dependencies are managed in the Poetry section below
dependencies = ["cosmpy (==0.9.3)", "dotenv (>=0.9.9,<0.10.0)", "black (>=25.1.0,<26.0.0)"]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[tool.poetry]
name = "fontana"
version = "0.1.0"
description = ""
authors = ["Nhu Viet Nguyen <braveryandglory@gmail.com>"]
readme = "README.md"
packages = [{include = "fontana", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
pytest = "^8.3.5"
pynacl = "^1.5.0"
# pylestia is managed as a git submodule, not as a package dependency
requests = "^2.32.3"
websockets = "^15.0.1"
ajsonrpc = "^1.2.0"
cosmpy = {version = "0.9.3", python = ">=3.13,<4.0"}

[tool.dotenv]
env_file = ".env"

[tool.poetry.scripts]
fontana = "fontana.cli.main:app"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
# The asyncio plugin can be configured with a marker instead
addopts = "--strict-markers"
markers = ["asyncio: mark test as using asyncio features"]
asyncio_mode = "strict"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
